Move PersistenceBeanTest into fat test folder
This required moving the methods in TestUtils, as they were used both by
the PersistenceBeanTest and other unit tests.

slurp() was moved to BasicChecks in test-utils, which makes it available
to all tests.

assertAssetList was removed and replaced with calls to assertThat()
using the ServerAssetByIdMatcher
